Protect user notes w auth middleware

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,14 @@ import { RequestHandler } from 'express';
 import NoteModel from '../models/note';
 import createHttpError from 'http-errors';
 import mongoose from 'mongoose';
+import { assertIsDefined } from '../util/assertIsDefined';
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const sessionUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(sessionUserId);
+    const notes = await NoteModel.find({ userId: sessionUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -13,9 +17,12 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 };
 
 export const getNote: RequestHandler = async (req, res, next) => {
+  const sessionUserId = req.session.userId;
   const noteId = req.params.noteId;
 
   try {
+    assertIsDefined(sessionUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note id.');
     }
@@ -23,6 +30,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
     const note = await NoteModel.findById(noteId).exec();
 
     if (!note) throw createHttpError(404, 'Note not found.');
+
+    if (!note.userId.equals(sessionUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
+    }
 
     res.status(200).json(note);
   } catch (error) {
@@ -41,12 +52,16 @@ export const createNote: RequestHandler<
   CreateNoteBody,
   unknown
 > = async (req, res, next) => {
+  const sessionUserId = req.session.userId;
   const title = req.body.title;
   const body = req.body.body;
 
   try {
+    assertIsDefined(sessionUserId);
+
     if (!title) throw createHttpError(400, 'Note must have a title.');
     const newNote = await NoteModel.create({
+      userId: sessionUserId,
       title,
       body,
     });
@@ -72,17 +87,28 @@ export const updateNote: RequestHandler<
   UpdateNoteBody,
   unknown
 > = async (req, res, next) => {
+  const sessionUserId = req.session.userId;
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newBody = req.body.body;
 
   try {
+    assertIsDefined(sessionUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note id.');
     }
+
     if (!newTitle) throw createHttpError(400, 'Note must have a title.');
+
     const note = await NoteModel.findById(noteId).exec();
+
     if (!note) throw createHttpError(404, 'Note not found.');
+
+    if (!note.userId.equals(sessionUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
+    }
+
     note.title = newTitle;
     note.body = newBody;
     const updatedNote = await note.save();
@@ -93,14 +119,25 @@ export const updateNote: RequestHandler<
 };
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
+  const sessionUserId = req.session.userId;
   const noteId = req.params.noteId;
+
   try {
+    assertIsDefined(sessionUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note id.');
     }
+
     const note = await NoteModel.findById(noteId).exec();
+
     if (!note) throw createHttpError(404, 'Note not found.');
-    await NoteModel.deleteOne({_id: noteId});
+
+    if (!note.userId.equals(sessionUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
+    }
+
+    await NoteModel.deleteOne({ _id: noteId }); // or note.remove()
     res.sendStatus(204);
   } catch (error) {
     next(error);
